Reverse java class
The given text file is read through a scanner. Each line in the text file is then stored into a stack called container.

Once the entire text file has been read through, a while loop keeps running until the stack is empty. Each loop, a line in the stack is popped. This first line popped should be the very last line in the text file. A for loop then starts at the end of the string and is then printed character by character back to the text file.

The end result is the entire text reversed, and each line reversed.